+++ pysrc/bytewax/metrics.py
@@ -1,5 +1,0 @@
-"""Interfaces for custom metrics."""
-
-from bytewax._bytewax import Gauge, Histogram
-
-__all__ = ["Gauge", "Histogram"]